+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component} from 'react';
 import './App.css';
-import Components from './Components/Random/RandomCards';
+import Components from './Components/Cards/NumberList';
 
 // TODO: Create a Header Component
 // App: The main component that handles all components for application.
